+++ src/routes/users.routes.js
@@ -1,24 +1,14 @@
 const express = require('express');
 const router = express.Router();
 const userController = require('../controllers/userController');
-const path = require('path');
-const multer = require('multer');
-const { body } = require('express-validator');
 
-const imgStorage = path.join(__dirname, '../../public/images/users');
-var mDStorage = multer.diskStorage({
-	destination: function (req, file, cb) {
-		cb(null, imgStorage);
-	},
-	filename: function (req, file, cb) {
-		cb(null, file.fieldname + '-' + Date.now() + path.extname(file.originalname));
-    }
-});
+const { body } = require('express-validator');
 
 
 // Middlewares
 const validations = require('../middlewares/validateRegisterMiddleware');
-const upload = multer({ storage: mDStorage});
+const upload = require('../middlewares/multerMiddleware');
+
 
 // Ruta hacía el login
 router.get('/login', userController.login);
